# internal/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Dockmesh API
  summary: Container management platform — HTTP API
  description: |
    REST API for Dockmesh. All endpoints are versioned under `/api/v1`.

    ## Authentication

    Two bearer-token flavours are accepted on the `Authorization` header:

    - **User JWTs** — obtained via `POST /auth/login`, short-lived (15 min),
      refreshable via `POST /auth/refresh`. Use for interactive UI / CLI sessions.
    - **API tokens** — prefix `dmt_`, long-lived, created under
      Settings → API Tokens. Use for CI, scripts, external integrations.

    Both share the same permission model — a token's role dictates what it can do.

    ## Versioning

    Version `v1` is stable in the sense that we don't remove fields or change
    response shapes without warning. New fields may be added at any time, so
    clients should ignore unknown fields gracefully.

    ## Errors

    Non-2xx responses use a uniform shape: `{"error": "<human-readable message>"}`.
    HTTP status carries the category (400/401/403/404/409/500).

    ## Maintenance

    This spec is hand-maintained. `TestOpenAPIDriftAgainstRoutes` compares the
    registered chi routes against the operations declared here and fails CI on
    mismatch — so every PR that touches a route must also touch this file.
  version: "1.0.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Dockmesh
    url: https://dockmesh.dev
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /containers/{id}:
    parameters:
      - $ref: "#/components/parameters/ContainerIDPath"
      - $ref: "#/components/parameters/HostQuery"
    get:
      operationId: containers.inspect
      tags: [containers]
      summary: Inspect one container
      responses:
        "200":
          description: Full ContainerJSON from the Docker daemon.
          content:
            application/json:
              schema: { type: object, description: Passthrough of the Docker SDK ContainerJSON. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: containers.remove
      tags: [containers]
      summary: Remove a container
      parameters:
        - name: force
          in: query
          schema: { type: boolean, default: false }
          description: Kill running containers before removing.
      responses:
        "204": { description: Removed. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /containers/{id}/restart:
    post:
      operationId: containers.restart
      tags: [containers]
      summary: Restart a container
      parameters:
        - $ref: "#/components/parameters/ContainerIDPath"
        - $ref: "#/components/parameters/HostQuery"
      responses:
        "204": { description: Restarted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /containers/{id}/unpause:
    post:
      operationId: containers.unpause
      tags: [containers]
      summary: Resume a paused container
      parameters:
        - $ref: "#/components/parameters/ContainerIDPath"
        - $ref: "#/components/parameters/HostQuery"
      responses:
        "204": { description: Resumed. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /images/{id}:
    get:
      operationId: images.inspect
      tags: [images]
      summary: Inspect an image
      description: |
        Returns the full Docker `ImageInspect` payload (RepoTags, RepoDigests,
        Config, RootFS layers, Architecture, Os, Created, Size, Labels, …).
        Remote agents currently respond 501 — image inspect is not yet
        carried over the agent protocol.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: Image id (sha256:…) or repo:tag.
        - $ref: "#/components/parameters/HostQuery"
      responses:
        "200":
          description: Docker ImageInspect.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "501":
          description: Inspect not yet supported on this host (remote agent).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      operationId: images.remove
      tags: [images]
      summary: Remove an image
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: Image id or name:tag.
        - $ref: "#/components/parameters/HostQuery"
        - name: force
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: List of layers / references deleted.
          content:
            application/json:
              schema:
                type: array
                items: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Image is in use by a container.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /volumes/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema: { type: string }
      - $ref: "#/components/parameters/HostQuery"
    get:
      operationId: volumes.inspect
      tags: [volumes]
      summary: Inspect one volume
      responses:
        "200":
          description: Volume.
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: volumes.remove
      tags: [volumes]
      summary: Remove a volume (must be unused unless forced)
      parameters:
        - name: force
          in: query
          schema: { type: boolean, default: false }
      responses:
        "204": { description: Removed. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Volume is in use.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /networks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
      - $ref: "#/components/parameters/HostQuery"
    get:
      operationId: networks.inspect
      tags: [networks]
      summary: Inspect one network
      responses:
        "200":
          description: Network.
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: networks.remove
      tags: [networks]
      summary: Remove a network
      responses:
        "204": { description: Removed. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Network still has active endpoints.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # -----------------------------------------------------------------------
  # backups — jobs, runs, targets (P.3.4)
  # -----------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "JWT or dmt_<40-char-base64url>"
      description: |
        Accepts either a user session JWT (from `POST /auth/login`) or an API
        token (prefix `dmt_`, from Settings → API Tokens). The route layer
        checks role permissions after authentication.

  # Schemas used across multiple groups live here. Group-specific schemas
  # live inline under the operation to keep the spec readable when new
  # groups get added.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid credentials
...
  parameters:
    HostQuery:
      name: host
      in: query
      description: "Target host id. Empty or `local` = central server's daemon; otherwise an agent UUID. `all` fans out to every online host."
      required: false
      schema:
        type: string
        default: local
...
    ContainerIDPath:
      name: id
      in: path
      description: Container id (long or short form) or name.
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but the token's role lacks the required permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...